pyproject.toml
[tool.poetry]
name = "resting"
version = "0.1.0"
description = ""
authors = ["Ahmed El-Zahaby <ahmedelzahaby8122000@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.2"
djangorestframework = "^3.14.0"
pyjwt = "^2.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
